Illegal move bug fix, material evaluation weights refactor
api.py/index.html changes:
------------------------
There was a bug in the client/server where when an illegal move was made, to restore the position,
the client would make a GET request to the /board endpoint which triggers another computer move.

To fix I created another endpoint /refresh which solely returns the game fen which is used by the client to restore the position.

engine.py changes:
-----------------
Material weights are now stored in a dictionary for the Evaluation class. When it comes time to tweak the engine, there wont be a need
to update both the _eval() method and the __repr__/__str__ method. These weights can be updated in the dict which will be reflected everywhere